# docs/swagger.yaml
definitions:
  dto.ArticleProgress:
    properties:
      article_id:
        example: 1
        type: integer
      completed:
        example: true
        type: boolean
      date_completed:
        example: "2023-08-15 14:30:00"
        type: string
    type: object
  dto.CompleteArticleRequest:
    properties:
      article_id:
        example: 1
        type: integer
      goal_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    required:
    - article_id
    - goal_id
    - user_id
    type: object
  dto.CreateGoalRequest:
    properties:
      articles_to_read:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      task:
        example: Study Go basics
        type: string
      title:
        example: Learn Golang
        type: string
    required:
    - task
    - title
    type: object
  dto.GoalProgressResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/dto.ArticleProgress'
        type: array
      completed:
        example: false
        type: boolean
      created_at:
        example: "2023-08-15 14:30:00"
        type: string
      id:
        example: 1
        type: integer
      task:
        example: Study Go basics
        type: string
      title:
        example: Learn Golang
        type: string
      total_articles:
        example: 3
        type: integer
      total_completed:
        example: 2
        type: integer
    type: object
  dto.GoalResponse:
    properties:
      articles_to_read:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      completed:
        example: false
        type: boolean
      created_at:
        example: "2023-08-15 14:30:00"
        type: string
      id:
        example: 1
        type: integer
      task:
        example: Study Go basics
        type: string
      title:
        example: Learn Golang
        type: string
    type: object
  dto.UpdateGoalRequest:
    properties:
      articles_to_read:
        example:
        - 4
        - 5
        - 6
        items:
          type: integer
        type: array
      completed:
        example: false
        type: boolean
      task:
        example: Study concurrency
        type: string
      title:
        example: Advanced Golang
        type: string
    required:
    - task
    - title
    type: object
info:
  contact: {}
paths:
  /pijar/goals/{user_id}:
    get:
      description: Retrieve all goals for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user goals
      tags:
      - Goals
    post:
      consumes:
      - application/json
      description: Add a new learning goal with articles to read
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Goal data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGoalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GoalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new goal
      tags:
      - Goals
  /pijar/goals/{user_id}/{id}:
    delete:
      description: Permanently delete a goal and its progress
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 'Contoh: {''message'':''Goal successfully deleted''}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a goal
      tags:
      - Goals
    put:
      consumes:
      - application/json
      description: Update goal details including articles to read
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GoalProgressResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a goal
      tags:
      - Goals
  /pijar/goals/complete-article:
    put:
      consumes:
      - application/json
      description: Mark an article as completed in a goal
      parameters:
      - description: Completion Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CompleteArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GoalProgressResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Complete an article progress
      tags:
      - Goals
swagger: "2.0"